Added support to show packages that are ahead of rosdistro
There are a couple of packages that are actually ahead of the rosdistro.
Partly this is because of premature updating, partly because once released
packages got downgraded again. Before this commit any version difference
showed as "outdated".

This commit adds proper version comparison operators for versions and uses them
to clearly separate outdated packages from packages being ahead.